backend/pyproject.toml
[tool.poetry]
name = "subtracker-backend"
version = "0.1.0"
description = ""
authors = ["luliangce <luliangce@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
django-ninja = "^0.16.2"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
